Bug 1380063 - Follow-up: show both buffer and local bookmarks at the top level.
This allows for folders, too.

There are two obstacles: one is that the factory is fixed and by the
time it gets involved it doesn’t know where you clicked; the other is
that `BookmarkFolder` plays two roles — as a rich `SQLiteBookmarkFolder`
it acts as the panel `source`, but plain ol’ `BookmarkFolder` is used as
a row item in the list itself.

So it’s not enough to try to track direction in `SQLiteBookmarkFolder`,
because at the moment it matters, we don’t actually have that
information!

And it’s hard to try to push it into `BookmarkFolder`, because those are
created by a factory from a cursor.

So the approach taken here is to give the factory the opportunity to
hand over a different factory in response to an index. Inside the
special factory we can decide whether the index points to local or to
buffer items.